Methods

Methods

This section describes systems methods implemented in Systems.jl.

Methods

Systems.statedim — Function.

statedim(s::AbstractSystem)

Returns the dimension of the state space of system s.

source

Systems.inputdim — Function.

inputdim(s::AbstractSystem)

Returns the dimension of the input space of system s.

source