Types

Types

This section describes systems types implemented in Systems.jl.

Types

Abstract Systems

Continuous Systems

Discrete Systems

Initial Value Problems

Abstract Systems

Systems.AbstractSystem — Type.

AbstractSystem

Abstract supertype for all system types.

source

Systems.AbstractContinuousSystem — Type.

AbstractContinuousSystem

Abstract supertype for all continuous system types.

source

Systems.AbstractDiscreteSystem — Type.

AbstractDiscreteSystem

Abstract supertype for all discrete system types.

source

Continuous Systems

Systems.ContinuousIdentitySystem — Type.

ContinuousIdentitySystem <: AbstractContinuousSystem

Trivial identity continuous-time system of the form

\[x' = 0.\]

source

Systems.ConstrainedContinuousIdentitySystem — Type.

ConstrainedContinuousIdentitySystem <: AbstractContinuousSystem

Trivial identity continuous-time system with state constraints of the form

\[x' = 0, x(t) ∈ \mathcal{X}.\]

source

Systems.LinearContinuousSystem — Type.

LinearContinuousSystem

Continuous-time linear system of the form

\[x' = A x.\]

Fields

A – square matrix

source

Systems.LinearControlContinuousSystem — Type.

LinearControlContinuousSystem

Continuous-time linear control system of the form

\[x' = A x + B u.\]

Fields

A – square matrix

B – matrix

source

Systems.ConstrainedLinearContinuousSystem — Type.

ConstrainedLinearContinuousSystem

Continuous-time linear system with state constraints of the form

\[x' = A x, x(t) ∈ \mathcal{X}.\]

Fields

A – square matrix

X – state constraints

source

Systems.ConstrainedLinearControlContinuousSystem — Type.

ConstrainedLinearControlContinuousSystem

Continuous-time linear control system with state constraints of the form

\[x' = A x + B u, x(t) ∈ \mathcal{X}, u(t) ∈ \mathcal{U} \text{ for all } t.\]

Fields

A – square matrix

B – matrix

X – state constraints

U – input constraints

source

Systems.LinearAlgebraicContinuousSystem — Type.

LinearAlgebraicContinuousSystem

Continuous-time linear algebraic system of the form

\[E x' = A x.\]

Fields

A – matrix

E – matrix, same size as A

source

Systems.ConstrainedLinearAlgebraicContinuousSystem — Type.

ConstrainedLinearAlgebraicContinuousSystem

Continuous-time linear system with state constraints of the form

\[E x' = A x, x(t) ∈ \mathcal{X}.\]

Fields

A – matrix

E – matrix, same size as A

X – state constraints

source

Discrete Systems

Systems.DiscreteIdentitySystem — Type.

DiscreteIdentitySystem <: AbstractDiscreteSystem

Trivial identity discrete-time system of the form

\[x_{k+1} = x_k.\]

source

Systems.ConstrainedDiscreteIdentitySystem — Type.

ConstrainedDiscreteIdentitySystem <: AbstractDiscreteSystem

Trivial identity discrete-time system with state constraints of the form

\[x_{k+1} = x_k, x_k ∈ \mathcal{X}.\]

source

Systems.LinearDiscreteSystem — Type.

LinearDiscreteSystem

Discrete-time linear system of the form

\[x_{k+1} = A x_k.\]

Fields

A – square matrix

source

Systems.LinearControlDiscreteSystem — Type.

LinearControlDiscreteSystem

Discrete-time linear control system of the form

\[x_{k+1} = A x_k + B u_k.\]

Fields

A – square matrix

B – matrix

source

Systems.ConstrainedLinearDiscreteSystem — Type.

ConstrainedLinearDiscreteSystem

Discrete-time linear system with state constraints of the form

\[x_{k+1} = A x_k, x_k ∈ \mathcal{X}.\]

Fields

A – square matrix

X – state constraints

source

Systems.ConstrainedLinearControlDiscreteSystem — Type.

ConstrainedLinearControlDiscreteSystem

Discrete-time linear control system with state constraints of the form

\[x_{k+1} = A x_k + B u_k, x_k ∈ \mathcal{X}, u_k ∈ \mathcal{U} \text{ for all } k.\]

Fields

A – square matrix

B – matrix

X – state constraints

U – input constraints

source

Systems.LinearAlgebraicDiscreteSystem — Type.

LinearAlgebraicDiscreteSystem

Discrete-time linear algebraic system of the form

\[E x_{k+1} = A x_k.\]

Fields

A – matrix

E – matrix, same size as A

source

Systems.ConstrainedLinearAlgebraicDiscreteSystem — Type.

ConstrainedLinearAlgebraicDiscreteSystem

Discrete-time linear system with state constraints of the form

\[E x_{k+1} = A x_k, x_k ∈ \mathcal{X}.\]

Fields

A – matrix

E – matrix, same size as A

X – state constraints

source

Initial Value Problems

Systems.InitialValueProblem — Type.

InitialValueProblem

Parametric composite type for initial value problems. It is parameterized in the system's type.

Fields

s – system

x0 – initial state

Examples

The linear system $x' = -x$ with initial condition $x_0 = [-1/2, 1/2]$:

julia> p = InitialValueProblem(LinearContinuousSystem(-eye(2)), [-1/2., 1/2]);

julia> p.x0
2-element Array{Float64,1}:
 -0.5
  0.5
julia> statedim(p)
  2
julia> inputdim(p)
  0

source

Systems.IVP — Type.

IVP

IVP is an alias for InitialValueProblem.

source