Internals

Internals

This section describes functions that are internal to the library.

Internals

Expression handling

Querying expressions

Expression handling

MathematicalSystems._corresponding_type — Function.

_corresponding_type(AT::Type{<:AbstractSystem}, fields::Tuple)

Return the system type whose field names match those in fields.

Input

AT – abstract system type, which can be either AbstractContinuousSystem or AbstractDiscreSystem

fields – tuple of field names

Output

The system type (either discrete or continous, depending on AT) whose fields names correspond to those in fields, or an error if the fields do not match any known system type.

Examples

julia> using MathematicalSystems: _corresponding_type

julia> _corresponding_type(AbstractContinuousSystem, ((:A),))
LinearContinuousSystem

julia> _corresponding_type(AbstractContinuousSystem, ((:A), (:B), (:X), (:U)))
ConstrainedLinearControlContinuousSystem

source

MathematicalSystems._capture_dim — Function.

_capture_dim(expr)

Return the tuple containing the dimension(s) in expr.

Input

expr – symbolic expression that can be of any of the following forms:

:x or :(x) – state dimension

:(x, u) – state and input dimension

:(x, u, w) – state, input and noise dimensions

Output

The scalar x if expr specifies the state dimension.

The vector [x, u] if expr specifies state and input dimension.

The vector [x, u, w] if expr specifies state, input and noise dimensions.

source

MathematicalSystems.add_asterisk — Function.

add_asterisk(summand, state::Symbol, input::Symbol, noise::Symbol)

Checks if expression summand contains state, input or noise at its end. If so, a multiplication expression, e.g. Expr(:call, :*, :A, :x) is created. If not,summand` is returned.

Input

summand – expressions

state – state variable

input – input variable

noise – noise variable, if available

Output

Multiplication expression or symbol.

Example

julia> using MathematicalSystems: add_asterisk

julia> add_asterisk(:(A1*x), :x, :u, :w)
:(A1 * x)

julia> add_asterisk(:(c1), :x, :u, :w)
:c1

julia> add_asterisk(:(Ax1), :x1, :u, :w)
:(A * x1)

julia> add_asterisk(:(Awb), :x1, :u, :wb)
:(A * wb)

julia> add_asterisk(:(A1u), :x, :u, :w)
:(A1 * u)

julia> add_asterisk(:(A1ub), :x, :u, :w)
:A1ub

source

Querying expressions

MathematicalSystems.is_equation — Function.

is_equation(expr)

Return true if the given expression expr corresponds to an equation lhs = rhs and false otherwise. This function just detects the presence of the symbol =.

Input

expr – expression

Output

A Bool indicating whether expr is an equation or not.

source

MathematicalSystems.extract_sum — Function.

extract_sum(summands, state::Symbol, input::Symbol, noise::Symbol)

Extract the variable name and field name for every element of summands which corresponds to the elements of the right-hand side of an affine system.

Input

summands – array of expressions

state – state variable

input – input variable

noise – noise variable

Output

Array of tuples of symbols with variable name and field name.

Example

julia> using MathematicalSystems: extract_sum

julia> extract_sum([:(A1*x)], :x, :u, :w)
1-element Array{Tuple{Any,Symbol},1}:
 (:A1, :A)

julia> extract_sum([:(A1*x), :(B1*u), :c], :x, :u, :w)
3-element Array{Tuple{Any,Symbol},1}:
 (:A1, :A)
 (:B1, :B)
 (:c, :c)

julia> extract_sum([:(A1*x7), :( B1*u7), :( B2*w7)], :x7, :u7, :w7)
3-element Array{Tuple{Any,Symbol},1}:
 (:A1, :A)
 (:B1, :B)
 (:B2, :D)

Notes

If an element of summands is a multiplication expression lhs*rhs, return lhs as variable name and :A as field name if rhs==state, :B as field name if rhs==input and :D as field name if rhs==noise.

If an element of summands is a symbol, and not equal to input or noise, the symbol is the variable name and the field name is :c. If it is equal to input, the variable name is a IdentityMultiple(I,state_dim) where state_dim is extracted from the state matrix (i.e. take the symbol lhs of lhs*rhs where rhs==state which corresponds to the state matrix and generate the expression state_dim = size(lhs,1) which is evaluated in the scope where @system is called) and the field name is :B.

Similiarily, if the element is equal to noise, the variable name is IdentityMultiple(I, state_dim) and the field name is :D.

source