Methods

Methods

This section describes systems methods implemented in MathematicalSystems.jl.

Methods

States

Inputs

Output

Traits

Maps

Successor

States

MathematicalSystems.statedim — Function.

statedim(s::AbstractSystem)

Returns the dimension of the state space of system s.

source

MathematicalSystems.stateset — Function.

stateset(s::AbstractSystem)

Returns the set of allowed states of system s.

source

Inputs

MathematicalSystems.inputdim — Function.

inputdim(s::AbstractSystem)

Returns the dimension of the input space of system s.

source

MathematicalSystems.inputset — Function.

inputset(s::AbstractSystem)

Returns the set of allowed inputs of system s.

source

MathematicalSystems.nextinput — Function.

nextinput(input::ConstantInput, n::Int=1)

Returns the first n elements of this input.

Input

input – a constant input

n – (optional, default: 1) the number of desired elements

Output

A repeated iterator that generates n equal samples of this input.

source

nextinput(input::VaryingInput, n::Int=1)

Returns the first n elements of this input.

Input

input – varying input

n – (optional, default: 1) number of desired elements

Output

An iterator of type Base.Iterators.Take that represents at most the first n elements of this input.

source

Output

MathematicalSystems.outputdim — Function.

outputdim(m::AbstractMap)

Returns the dimension of the output space of the map m.

source

MathematicalSystems.outputmap — Function.

outputmap(s::SystemWithOutput)

Returns the output map of a system with output.

source

Traits

MathematicalSystems.islinear — Method.

islinear(s::AbstractSystem)

Specifies if the dynamics of system s is specified by linear equations.

Notes

We adopt the notion from [Section 2.7, 1]. For example, the system with inputs $x' = f(t, x, u) = A x + B u$ is linear, since the function $f(t, ⋅, ⋅)$ is linear in $(x, u)$ for each $t ∈ \mathbb{R}$. On the other hand, $x' = f(t, x, u) = A x + B u + c$ is affine but not linear, since it is not linear in $(x, u)$.

This function uses the information of the type, not the value. So, if a system type allows an instance that is not linear, it returns false by default. For example, polynomial systems can be nonlinear; hence islinear returns false.

[1] Sontag, Eduardo D. Mathematical control theory: deterministic finite dimensional systems. Vol. 6. Springer Science & Business Media, 2013.

source

MathematicalSystems.islinear — Method.

islinear(m::AbstractMap)

Specifies if the map m is linear or not.

Notes

A map is linear if it preserves the operations of scalar multiplication and vector addition.

source

MathematicalSystems.isaffine — Method.

isaffine(s::AbstractSystem)

Specifies if the dynamics of system s is specified by affine equations.

Notes

An affine system is the composition of a linear system and a translation. See islinear(::AbstractSystem) for the notion of linear system adopted in this library.

This function uses the information of the type, not the value. So, if a system type allows an instance that is not affine, it returns false by default. For example, polynomial systems can be nonlinear; hence isaffine is false.

source

MathematicalSystems.ispolynomial — Method.

ispolynomial(s::AbstractSystem)

Specifies if the dynamics of system s is specified by polynomial equations.

The criterion refers to the type information, not the value. Hence, e.g. a LinearContinuousSystem is not considered to be of polynomial type.

source

MathematicalSystems.isaffine — Method.

isaffine(m::AbstractMap)

Specifies if the map m is affine or not.

Notes

An affine map is the composition of a linear map and a translation. See also islinear(::AbstractMap).

source

Maps

MathematicalSystems.apply — Function.

apply(m::AbstractMap, args...)

Apply the rule specified by the map to the given arguments.

source

Successor

MathematicalSystems.successor — Function.

successor(system::DiscreteIdentitySystem, x::AbstractVector)

Return the successor state of a DiscreteIdentitySystem.

Input

system – DiscreteIdentitySystem

x – state (it should be any vector type)

Output

The same state x.

source

successor(system::ConstrainedDiscreteIdentitySystem, x::AbstractVector;
          [check_constraints]=true)

Return the successor state of a ConstrainedDiscreteIdentitySystem.

Input

system – ConstrainedDiscreteIdentitySystem

x – state (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The same state x.

source

successor(system::LinearDiscreteSystem, x::AbstractVector)

Return the successor state of a LinearDiscreteSystem.

Input

system – LinearDiscreteSystem

x – state (it should be any vector type)

Output

The result of applying the system to x.

source

successor(system::AffineDiscreteSystem, x::AbstractVector)

Return the successor state of a AffineDiscreteSystem.

Input

system – AffineDiscreteSystem

x – state (it should be any vector type)

Output

The result of applying the system to x.

source

successor(system::LinearControlDiscreteSystem, x::AbstractVector, u::AbstractVector)

Return the successor state of a LinearControlDiscreteSystem.

Input

system – LinearControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

Output

The result of applying the system to x, with input u.

source

successor(system::ConstrainedLinearDiscreteSystem, x::AbstractVector;
          [check_constraints]=true)

Return the successor state of a ConstrainedLinearDiscreteSystem.

Input

system – ConstrainedLinearDiscreteSystem

x – state (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x.

source

successor(system::ConstrainedAffineDiscreteSystem, x::AbstractVector;
          [check_constraints])

Return the successor state of a ConstrainedAffineDiscreteSystem.

Input

system – ConstrainedAffineDiscreteSystem

x – state (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x.

source

successor(system::ConstrainedLinearControlDiscreteSystem, x::AbstractVector,
          u::AbstractVector; [check_constraints]=true)

Return the successor state of a ConstrainedLinearControlDiscreteSystem.

Input

system – ConstrainedLinearControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

check_constraints – (optional, default: true) check if the state (resp. input) belongs to the state set (resp. input set)

Output

The result of applying the system to x, with input u.

source

successor(system::ConstrainedAffineControlDiscreteSystem, x::AbstractVector,
          u::AbstractVector; [check_constraints]=true)

Return the successor state of a ConstrainedAffineControlDiscreteSystem.

Input

system – ConstrainedAffineControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

check_constraints – (optional, default: true) check if the state (resp. input) belongs to the state set (resp. input set)

Output

The result of applying the system to x, with input u.

source

successor(system::BlackBoxDiscreteSystem, x::AbstractVector)

Return the successor state of a BlackBoxDiscreteSystem.

Input

system – BlackBoxDiscreteSystem

x – state (it should be any vector type)

Output

The result of applying the system to x.

source

successor(system::ConstrainedBlackBoxDiscreteSystem, x::AbstractVector;
          [check_constraints]=true)

Return the successor state of a ConstrainedBlackBoxDiscreteSystem.

Input

system – ConstrainedBlackBoxDiscreteSystem

x – state (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x.

source

successor(system::ConstrainedBlackBoxControlDiscreteSystem, x::AbstractVector,
          u::AbstractVector; [check_constraints]=true)

Return the successor state of a ConstrainedBlackBoxControlDiscreteSystem.

Input

system – ConstrainedBlackBoxControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x, with input u.

source

successor(system::NoisyConstrainedLinearDiscreteSystem, x::AbstractVector,
          w::AbstractVector; [check_constraints]=true)

Return the successor state of a NoisyConstrainedLinearDiscreteSystem.

Input

system – NoisyConstrainedLinearDiscreteSystem

x – state (it should be any vector type)

w – noise (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x, with noise w.

source

successor(system::NoisyConstrainedLinearControlDiscreteSystem,
          x::AbstractVector, u::AbstractVector, w::AbstractVector; [check_constraints]=true)

Return the successor state of a NoisyConstrainedLinearControlDiscreteSystem.

Input

system – NoisyConstrainedLinearControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

w – noise (it should be any vector type)

check_constraints – (optional, default: true) check if the state (resp. input) belongs to the state set (resp. input set)

Output

The result of applying the system to x, with input u and noise w.

source

successor(system::NoisyConstrainedAffineControlDiscreteSystem, x::AbstractVector,
          u::AbstractVector; [check_constraints]=true)

Return the successor state of a NoisyConstrainedAffineControlDiscreteSystem.

Input

system – NoisyConstrainedAffineControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

w – noise (it should be any vector type)

check_constraints – (optional, default: true) check if the state (resp. input) belongs to the state set (resp. input set)

Output

The result of applying the system to x, with input u and noise w.

source

successor(system::NoisyConstrainedBlackBoxControlDiscreteSystem, x::AbstractVector,
          u::AbstractVector, w::AbstractVector; [check_constraints]=true)

Return the successor state of a NoisyConstrainedBlackBoxControlDiscreteSystem.

Input

system – NoisyConstrainedBlackBoxControlDiscreteSystem

x – state (it should be any vector type)

u – input (it should be any vector type)

w – noise (it should be any vector type)

check_constraints – (optional, default: true) check if the state belongs to the state set

Output

The result of applying the system to x, with input u and noise w.

source