Methods

Methods

This section describes systems methods implemented in MathematicalSystems.jl.

Methods

States

Inputs

Output

States

MathematicalSystems.statedim — Function.

statedim(s::AbstractSystem)

Returns the dimension of the state space of system s.

source

MathematicalSystems.stateset — Function.

stateset(s::AbstractSystem)

Returns the set of allowed states of system s.

source

Inputs

MathematicalSystems.inputdim — Function.

inputdim(s::AbstractSystem)

Returns the dimension of the input space of system s.

source

MathematicalSystems.inputset — Function.

inputset(s::AbstractSystem)

Returns the set of allowed inputs of system s.

source

MathematicalSystems.nextinput — Function.

nextinput(input::ConstantInput, n::Int=1)

Returns the first n elements of this input.

Input

input – a constant input

n – (optional, default: 1) the number of desired elements

Output

A repeated iterator that generates n equal samples of this input.

source

nextinput(input::VaryingInput, n::Int=1)

Returns the first n elements of this input.

Input

input – varying input

n – (optional, default: 1) number of desired elements

Output

An iterator of type Base.Iterators.Take that represents at most the first n elements of this input.

source

Output

MathematicalSystems.outputdim — Function.

outputdim(m::AbstractMap)

Returns the dimension of the output space of the map m.

source

MathematicalSystems.outputmap — Function.

outputmap(s::SystemWithOutput)

Returns the output map of a system with output.

source