Types

Types

This section describes systems types implemented in Systems.jl.

Types

Abstract Systems

Continuous Systems

Discrete Systems

Initial Value Problems

Input Types

Abstract Systems

Systems.AbstractSystem — Type.

AbstractSystem

Abstract supertype for all system types.

source

Systems.AbstractContinuousSystem — Type.

AbstractContinuousSystem

Abstract supertype for all continuous system types.

source

Systems.AbstractDiscreteSystem — Type.

AbstractDiscreteSystem

Abstract supertype for all discrete system types.

source

Continuous Systems

Systems.ContinuousIdentitySystem — Type.

ContinuousIdentitySystem <: AbstractContinuousSystem

Trivial identity continuous-time system of the form

\[x' = 0.\]

source

Systems.ConstrainedContinuousIdentitySystem — Type.

ConstrainedContinuousIdentitySystem <: AbstractContinuousSystem

Trivial identity continuous-time system with state constraints of the form

\[x' = 0, x(t) ∈ \mathcal{X}.\]

source

Systems.LinearContinuousSystem — Type.

LinearContinuousSystem

Continuous-time linear system of the form

\[x' = A x.\]

Fields

A – square matrix

source

Systems.LinearControlContinuousSystem — Type.

LinearControlContinuousSystem

Continuous-time linear control system of the form

\[x' = A x + B u.\]

Fields

A – square matrix

B – matrix

source

Systems.ConstrainedLinearContinuousSystem — Type.

ConstrainedLinearContinuousSystem

Continuous-time linear system with state constraints of the form

\[x' = A x, x(t) ∈ \mathcal{X}.\]

Fields

A – square matrix

X – state constraints

source

Systems.ConstrainedLinearControlContinuousSystem — Type.

ConstrainedLinearControlContinuousSystem

Continuous-time linear control system with state constraints of the form

\[x' = A x + B u, x(t) ∈ \mathcal{X}, u(t) ∈ \mathcal{U} \text{ for all } t.\]

Fields

A – square matrix

B – matrix

X – state constraints

U – input constraints

source

Systems.LinearAlgebraicContinuousSystem — Type.

LinearAlgebraicContinuousSystem

Continuous-time linear algebraic system of the form

\[E x' = A x.\]

Fields

A – matrix

E – matrix, same size as A

source

Systems.ConstrainedLinearAlgebraicContinuousSystem — Type.

ConstrainedLinearAlgebraicContinuousSystem

Continuous-time linear system with state constraints of the form

\[E x' = A x, x(t) ∈ \mathcal{X}.\]

Fields

A – matrix

E – matrix, same size as A

X – state constraints

source

Discrete Systems

Systems.DiscreteIdentitySystem — Type.

DiscreteIdentitySystem <: AbstractDiscreteSystem

Trivial identity discrete-time system of the form

\[x_{k+1} = x_k.\]

source

Systems.ConstrainedDiscreteIdentitySystem — Type.

ConstrainedDiscreteIdentitySystem <: AbstractDiscreteSystem

Trivial identity discrete-time system with state constraints of the form

\[x_{k+1} = x_k, x_k ∈ \mathcal{X}.\]

source

Systems.LinearDiscreteSystem — Type.

LinearDiscreteSystem

Discrete-time linear system of the form

\[x_{k+1} = A x_k.\]

Fields

A – square matrix

source

Systems.LinearControlDiscreteSystem — Type.

LinearControlDiscreteSystem

Discrete-time linear control system of the form

\[x_{k+1} = A x_k + B u_k.\]

Fields

A – square matrix

B – matrix

source

Systems.ConstrainedLinearDiscreteSystem — Type.

ConstrainedLinearDiscreteSystem

Discrete-time linear system with state constraints of the form

\[x_{k+1} = A x_k, x_k ∈ \mathcal{X}.\]

Fields

A – square matrix

X – state constraints

source

Systems.ConstrainedLinearControlDiscreteSystem — Type.

ConstrainedLinearControlDiscreteSystem

Discrete-time linear control system with state constraints of the form

\[x_{k+1} = A x_k + B u_k, x_k ∈ \mathcal{X}, u_k ∈ \mathcal{U} \text{ for all } k.\]

Fields

A – square matrix

B – matrix

X – state constraints

U – input constraints

source

Systems.LinearAlgebraicDiscreteSystem — Type.

LinearAlgebraicDiscreteSystem

Discrete-time linear algebraic system of the form

\[E x_{k+1} = A x_k.\]

Fields

A – matrix

E – matrix, same size as A

source

Systems.ConstrainedLinearAlgebraicDiscreteSystem — Type.

ConstrainedLinearAlgebraicDiscreteSystem

Discrete-time linear system with state constraints of the form

\[E x_{k+1} = A x_k, x_k ∈ \mathcal{X}.\]

Fields

A – matrix

E – matrix, same size as A

X – state constraints

source

Initial Value Problems

Systems.InitialValueProblem — Type.

InitialValueProblem

Parametric composite type for initial value problems. It is parameterized in the system's type.

Fields

s – system

x0 – initial state

Examples

The linear system $x' = -x$ with initial condition $x_0 = [-1/2, 1/2]$:

julia> p = InitialValueProblem(LinearContinuousSystem(-eye(2)), [-1/2., 1/2]);

julia> p.x0
2-element Array{Float64,1}:
 -0.5
  0.5
julia> statedim(p)
  2
julia> inputdim(p)
  0

source

Systems.IVP — Type.

IVP

IVP is an alias for InitialValueProblem.

source

Input Types

Systems.AbstractInput — Type.

AbstractInput

Abstract supertype for all input types.

Notes

The input types defined here implement an iterator interface, such that other methods can build upon the behavior of inputs which are either constant or varying.

Iteration is supported with an index number called iterator state. The iteration function Base.next takes and returns a tuple (input, state), where input represents the value of the input, and state is an index which counts the number of times this iterator was called.

A convenience function nextinput(input, n) is also provided and it returns the first n elements of input.

source

Systems.ConstantInput — Type.

ConstantInput{UT} <: AbstractInput

Type representing an input that remains constant in time.

Fields

U – input set

Examples

The constant input holds a single element and its length is infinite. To access the field U, you can use Base's next given a state, or the method nextinput given the number of desired input elements:

julia> c = ConstantInput(-1//2)
Systems.ConstantInput{Rational{Int64}}(-1//2)

julia> next(c, 1)
(-1//2, nothing)

julia> next(c, 2)
(-1//2, nothing)

julia> collect(nextinput(c, 4))
4-element Array{Rational{Int64},1}:
 -1//2
 -1//2
 -1//2
 -1//2

The elements of this input are rational numbers:

julia> eltype(c)
Rational{Int64}

To transform a constant input, you can use map as in:

julia> map(x->2*x, c)
Systems.ConstantInput{Rational{Int64}}(-1//1)

source

Systems.VaryingInput — Type.

VaryingInput{UT} <: AbstractInput

Type representing an input that may vary with time.

Fields

U – vector of input sets

Examples

The varying input holds a vector and its length equals the number of elements in the vector. Consider an input given by a vector of rational numbers:

julia> v = VaryingInput([-1//2, 1//2])
Systems.VaryingInput{Rational{Int64}}(Rational{Int64}[-1//2, 1//2])

julia> length(v)
2

julia> eltype(v)
Rational{Int64}

Base's next method receives the input and an integer state and returns the input element and the next iteration state:

julia> next(v, 1)
(-1//2, 2)

julia> next(v, 2)
(1//2, 3)

The method nextinput receives a varying input and an integer n and returns an iterator over the first n elements of this input (where n=1 by default):

julia> typeof(nextinput(v))
Base.Iterators.Take{Systems.VaryingInput{Rational{Int64}}}

julia> collect(nextinput(v, 1))
1-element Array{Rational{Int64},1}:
 -1//2

julia> collect(nextinput(v, 2))
2-element Array{Rational{Int64},1}:
 -1//2
  1//2

You can collect the inputs in an array, or equivalently use list comprehension, (or use a for loop):

julia> collect(v)
2-element Array{Rational{Int64},1}:
 -1//2
  1//2

julia> [2*vi for vi in v]
2-element Array{Rational{Int64},1}:
 -1//1
  1//1

Since this input type is finite, querying more elements than its length returns the full vector:

julia> collect(nextinput(v, 4))
2-element Array{Rational{Int64},1}:
 -1//2
  1//2

To transform a varying input, you can use map as in:

julia> map(x->2*x, v)
Systems.VaryingInput{Rational{Int64}}(Rational{Int64}[-1//1, 1//1])

source